refactoring: CommonModal & custom hook
CommonModal 폴더로 이동
