+++ frontend/guest-app/src/components/CommonModal/CommonModal.js
