+++ frontend/guest-app/src/components/Modals/EditPriofileModal/EditProfileModal.js
@@ -2,7 +2,7 @@ import React, {useState} from "react";
 import {Grid, Typography} from "@material-ui/core";
 import Button from "@material-ui/core/Button";
 import UserAvatar from "../../UserAvatar/UserAvatar.js";
-import CommonModal from "../../CommonModal.js";
+import CommonModal from "../../CommonModal/CommonModal.js";
 import TextInput from "./TextInput.js";
 import {CompanyIcon, EmailIcon, UserIcon} from "../../FontAwesomeIcons.js";
 import useTextInput from "./useTextInput.js";
