+++ frontend/guest-app/src/components/Modals/LogoutModal.js
@@ -1,7 +1,7 @@
 import React from "react";
 import {Grid} from "@material-ui/core";
 import Button from "@material-ui/core/Button";
-import CommonModal from "../CommonModal.js";
+import CommonModal from "../CommonModal/CommonModal.js";
 
 function LogOutModal({isOpened = false, onCancelClick, onLogout}) {
 	return (
