+++ frontend/guest-app/src/components/SideMenu/SideMenuBody.js
@@ -5,7 +5,7 @@ import SideMenuItem from "./SideMenuItem.js";
 import EditProfileModal from "../Modals/EditPriofileModal/EditProfileModal.js";
 import useSideMenuStyles from "./UseSideMenuStyles.js";
 import {LogoutIcon, QuestionIcon, UserIcon} from "../FontAwesomeIcons.js";
-import useCommonModal from "../useCommonModal.js";
+import useCommonModal from "../CommonModal/useCommonModal.js";
 import LogOutModal from "../Modals/LogoutModal.js";
 import MyQuestionModal from "../Modals/MyQuestionModal.js";
